+++ src/ExecuteStage.h
@@ -15,15 +15,15 @@ struct EControlSignals{
 	bool MemWriteEn;
 	bool MemReadEn;
 	bool ALUsrc;
-	uint8_t ALUOp;
-	uint8_t RegDst;
+	int32_t ALUOp;
+	int32_t RegDst;
 	bool JAL;
-	uint32_t readdata1;
-	uint32_t readdata2;
-	uint32_t immediate;
-	uint8_t rs;
-	uint8_t rt;
-	uint8_t rd;
+	int32_t readdata1;
+	int32_t readdata2;
+	int32_t immediate;
+	int32_t rs;
+	int32_t rt;
+	int32_t rd;
 };
 
 
@@ -46,34 +46,34 @@ private:
 private:
 	//local stage needs: 
 	std::thread Executethread;
-	uint32_t PC;				//Programcounter
-	uint32_t MC;				//MachineCode
+	int32_t PC;				//Programcounter
+	int32_t MC;				//MachineCode
 	bool running = true; // temporary for debugging purposes.(so we dont use 100%CPU)
 	EControlSignals EXEdata;
-	uint32_t Operand1, Operand2;
-	uint32_t Out_JALM, Out_BOP2M;
-	uint8_t Out_RegDstM;
+	int32_t Operand1, Operand2;
+	int32_t Out_JALM, Out_BOP2M;
+	int32_t Out_RegDstM;
 
 private: //functions 
 	void Executejob();
-	uint32_t ALU(uint32_t operand1, uint32_t operand2, uint8_t opSel);
-	void JalMux(uint32_t PC, uint32_t Rs, bool JalSignal);
-	void Op1Mux(uint32_t JalMux, uint32_t WB32, uint32_t MEM32, uint8_t ForwardA);
-	void BeforeOp2Mux(uint32_t Rt, uint32_t WB32, uint32_t MEM32, uint8_t ForwardB);
-	void Op2Mux(uint32_t BOP2Mux, uint32_t Imm, uint8_t AluSrc);
-	void RegDstMux(uint8_t rt, uint8_t rd, uint8_t RegDstS);
+	int32_t ALU(int32_t operand1, int32_t operand2, int32_t opSel);
+	void JalMux(int32_t PC, int32_t Rs, bool JalSignal);
+	void Op1Mux(int32_t JalMux, int32_t WB32, int32_t MEM32, int32_t ForwardA);
+	void BeforeOp2Mux(int32_t Rt, int32_t WB32, int32_t MEM32, int32_t ForwardB);
+	void Op2Mux(int32_t BOP2Mux, int32_t Imm, int32_t AluSrc);
+	void RegDstMux(int32_t rt, int32_t rd, int32_t RegDstS);
 
 private:
-	static constexpr uint8_t _ADD = 0b0000;
-	static constexpr uint8_t _SUB = 0b0001;
-	static constexpr uint8_t _AND = 0b0010;
-	static constexpr uint8_t _OR = 0b0011;
-	static constexpr uint8_t _NOR = 0b0100;
-	static constexpr uint8_t _XOR = 0b0101;
-	static constexpr uint8_t _SLT = 0b0110;
-	static constexpr uint8_t _SLL = 0b0111;
-	static constexpr uint8_t _SRL = 0b1000;
-	static constexpr uint8_t _SGT = 0b1001;
+	static constexpr int32_t _ADD = 0b0000;
+	static constexpr int32_t _SUB = 0b0001;
+	static constexpr int32_t _AND = 0b0010;
+	static constexpr int32_t _OR = 0b0011;
+	static constexpr int32_t _NOR = 0b0100;
+	static constexpr int32_t _XOR = 0b0101;
+	static constexpr int32_t _SLT = 0b0110;
+	static constexpr int32_t _SLL = 0b0111;
+	static constexpr int32_t _SRL = 0b1000;
+	static constexpr int32_t _SGT = 0b1001;
 
 public:
