+++ src/ForwardingUnit.cpp
@@ -29,7 +29,7 @@ void ForwardingUnit::evaluateForwarding()
     }
 }
 
-void ForwardingUnit::FUinputEXE(uint8_t EXErs, uint8_t EXErt) {
+void ForwardingUnit::FUinputEXE(int32_t EXErs, int32_t EXErt) {
     if (this == nullptr) {
         std::cerr << "Error: Attempted to call FUinputEXE on a null pointer!" << std::endl;
         return;
@@ -37,14 +37,14 @@ void ForwardingUnit::FUinputEXE(uint8_t EXErs, uint8_t EXErt) {
     this->IDEXE_Rs = EXErs;
     this->IDEXE_Rt = EXErt;
 }
-void ForwardingUnit::FUinputMEM(bool EXEMEM_RegWrite, uint8_t EXEMEM_Rd, uint32_t MEMdata) {
+void ForwardingUnit::FUinputMEM(bool EXEMEM_RegWrite, int32_t EXEMEM_Rd, int32_t MEMdata) {
     this->MEMdata = MEMdata; //this has nothing to do with Forwarding Logic, Its just to pass data to the Alu Mux.
                              //##Can be moved to another function for better clarity,But requires further syncronization,##
     this->EXEMEM_RegWrite = EXEMEM_RegWrite;
     this->EXEMEM_Rd = EXEMEM_Rd;
     MEMsemaphore.release();
 }
-void ForwardingUnit::FUinputWB(bool MEMWB_RegWrite, uint8_t MEMWB_Rd, uint32_t WBdata) {
+void ForwardingUnit::FUinputWB(bool MEMWB_RegWrite, int32_t MEMWB_Rd, int32_t WBdata) {
     this->WBdata = WBdata;//this has nothing to do with Forwarding Logic, Its just to pass data to the Alu Mux.
                           //##Can be moved to another function for better clarity,But requires further syncronization,##
     this->MEMWB_RegWrite = MEMWB_RegWrite;
