+++ src/IDEXE.h
@@ -13,39 +13,39 @@ class IDEXE {
 private:
 	std::binary_semaphore s1;  // Binary semaphore for raw
 	
-	uint32_t PC = 0x00000000;	//Programcounter
-	uint32_t MC = 0x00000000; //MachineCode
-	uint8_t ALUOp=0x0; 
-	uint8_t RegDst=0x0;
+	int32_t PC = 0x00000000;	//Programcounter
+	int32_t MC = 0x00000000; //MachineCode
+	int32_t ALUOp=0x0;
+	int32_t RegDst=0x0;
 	bool JAL=false;
 	bool ALUsrc= false;
 	bool MemReadEn = false;
 	bool MemWriteEn= false;
 	bool MemtoReg  = false;
 	bool RegWriteEn= false;
-	uint32_t readdata1=0x0;
-	uint32_t readdata2= 0x0;
-	uint32_t immediate= 0x0;
-	uint8_t rs= 0x0;
-	uint8_t rt= 0x0;
-	uint8_t rd= 0x0;
+	int32_t readdata1=0x0;
+	int32_t readdata2= 0x0;
+	int32_t immediate= 0x0;
+	int32_t rs= 0x0;
+	int32_t rt= 0x0;
+	int32_t rd= 0x0;
 	
 	
 public : 
 	IDEXE();
 	void writedata(bool RegWriteEnin, bool MemtoRegin, bool MemWriteEnin, bool MemReadEnin,
-		bool ALUsrcin, uint8_t ALUOpin, uint8_t RegDstin,bool JALin,
-		uint32_t PCin, uint32_t MC,
-		uint32_t readdata1in, uint32_t readdata2in,
-		uint32_t immediatein,
-		uint8_t rsin, uint8_t rtin, uint8_t rdin);
+		bool ALUsrcin, int32_t ALUOpin, int32_t RegDstin,bool JALin,
+		int32_t PCin, int32_t MC,
+		int32_t readdata1in, int32_t readdata2in,
+		int32_t immediatein,
+		int32_t rsin, int32_t rtin, int32_t rdin);
 
 	void readdata(bool& MemWriteEnOut, bool& MemtoRegOut, bool& RegWriteEnOut, bool& MemReadEnOut,
-		bool& ALUsrcOut, uint8_t& ALUOpOut, uint8_t& RegDstOut, bool& JALOut,
-		uint32_t& PCOut, uint32_t& MCOut,
-		uint32_t& readdata1Out, uint32_t& readdata2Out,
-		uint32_t& immediateOut,
-		uint8_t& rsOut, uint8_t& rtOut, uint8_t& rdOut);
+		bool& ALUsrcOut, int32_t& ALUOpOut, int32_t& RegDstOut, bool& JALOut,
+		int32_t& PCOut, int32_t& MCOut,
+		int32_t& readdata1Out, int32_t& readdata2Out,
+		int32_t& immediateOut,
+		int32_t& rsOut, int32_t& rtOut, int32_t& rdOut);
 };
 
 #endif
