+++ src/RegisterFile.h
@@ -11,7 +11,7 @@ class RegisterFile {
 private:
 
 
-    std::array<uint32_t, 32> registers;  // 32 general-purpose registers
+    std::array<int32_t, 32> registers;  // 32 general-purpose registers
     std::binary_semaphore reg_semaphore; // Semaphore for later sync
 
     
@@ -22,10 +22,10 @@ public:
     RegisterFile();
 
     // Read two registers (source operands)
-    void readRegisters(uint8_t readReg1, uint8_t readReg2, uint32_t& data1, uint32_t& data2);
+    void readRegisters(int32_t readReg1, int32_t readReg2, int32_t& data1, int32_t& data2);
 
     // Write to a register (destination register)
-    void writeRegister(uint8_t writeReg, uint32_t writeData, bool regWrite);
+    void writeRegister(int32_t writeReg, int32_t writeData, bool regWrite);
 
     // Reset all registers to zero and initialize SP and PC
     void reset();
